Extract public file. Modify statistics table location
