+++ src/views/components/common/trace-chart-table/trace-statistics-item.vue
