+++ src/views/components/common/trace-chart-table/trace.scss
@@ -40,3 +40,21 @@
   width: 150px;
   text-align: center;
 }
+.max-time {
+  width: 150px;
+}
+.method {
+  width: 300px;
+}
+.avg-time {
+  width: 150px;
+}
+.min-time {
+  width: 150px;
+}
+.count {
+  width: 120px;
+}
+.sum-time {
+  width: 150px;
+}
